Rename back to constants
