+++ conda_lockfiles/loaders/pixi.py
@@ -9,7 +9,7 @@ from conda.base.context import context
 from conda.models.records import PackageRecord
 from ruamel.yaml import YAML
 
-from ..formats import PIXI_LOCK_FILE
+from ..constants import PIXI_LOCK_FILE
 from .base import BaseLoader, build_number_from_build_string
 
 if TYPE_CHECKING:
