+++ tests/loaders/test_explicit.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from conda_lockfiles.formats import EXPLICIT_KEY
+from conda_lockfiles.constants import EXPLICIT_KEY
 from conda_lockfiles.loaders.explicit import ExplicitLoader
 
 from .. import EXPLICIT_LOCK_DIR
